+++ org.hl7.fhir.validation/src/main/java/org/hl7/fhir/validation/cli/services/StandAloneValidatorFetcher.java
@@ -1,16 +1,8 @@
 package org.hl7.fhir.validation.cli.services;
 
-import java.io.IOException;
-import java.net.MalformedURLException;
-import java.net.URISyntaxException;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Locale;
-
+import com.google.gson.JsonObject;
 import org.hl7.fhir.convertors.txClient.TerminologyClientFactory;
-import org.hl7.fhir.exceptions.DefinitionException;
 import org.hl7.fhir.exceptions.FHIRException;
-import org.hl7.fhir.exceptions.FHIRFormatError;
 import org.hl7.fhir.r5.context.IWorkerContext;
 import org.hl7.fhir.r5.elementmodel.Element;
 import org.hl7.fhir.r5.model.CanonicalResource;
@@ -22,26 +14,23 @@ import org.hl7.fhir.utilities.VersionUtilities;
 import org.hl7.fhir.utilities.VersionUtilities.VersionURLInfo;
 import org.hl7.fhir.utilities.json.JSONUtil;
 import org.hl7.fhir.utilities.json.JsonTrackingParser;
-import org.hl7.fhir.utilities.npm.BasePackageCacheManager;
 import org.hl7.fhir.utilities.npm.FilesystemPackageCacheManager;
 import org.hl7.fhir.utilities.npm.NpmPackage;
 
-import com.google.gson.JsonObject;
-
-import ca.uhn.fhir.util.JsonUtil;
+import java.io.IOException;
+import java.net.MalformedURLException;
+import java.net.URISyntaxException;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Locale;
 
 public class StandAloneValidatorFetcher implements IValidatorResourceFetcher {
-
-  public interface IPackageInstaller {
-    boolean packageExists(String id, String ver) throws IOException, FHIRException;
-    void loadPackage(String id, String ver) throws IOException, FHIRException;
-  }
 
   List<String> mappingsUris = new ArrayList<>();
   private FilesystemPackageCacheManager pcm;
   private IWorkerContext context;
   private IPackageInstaller installer;
-  
+
   public StandAloneValidatorFetcher(FilesystemPackageCacheManager pcm, IWorkerContext context, IPackageInstaller installer) {
     super();
     this.pcm = pcm;
@@ -50,8 +39,8 @@ public class StandAloneValidatorFetcher implements IValidatorResourceFetcher {
   }
 
   @Override
-  public Element fetch(Object appContext, String url) throws FHIRFormatError, DefinitionException, FHIRException, IOException {
-    throw new FHIRException("The URL '"+url+"' is not known to the FHIR validator, and has not been provided as part of the setup / parameters");
+  public Element fetch(Object appContext, String url) throws FHIRException {
+    throw new FHIRException("The URL '" + url + "' is not known to the FHIR validator, and has not been provided as part of the setup / parameters");
   }
 
   @Override
@@ -64,15 +53,15 @@ public class StandAloneValidatorFetcher implements IValidatorResourceFetcher {
     if (!Utilities.isAbsoluteUrl(url)) {
       return false;
     }
-    
+
     if (url.contains("|")) {
       url = url.substring(0, url.lastIndexOf("|"));
     }
-    
+
     if (type.equals("uri") && isMappingUri(url)) {
       return true;
     }
-    
+
     // if we've got to here, it's a reference to a FHIR URL. We're going to try to resolve it on the fly
     String pid = null;
     String ver = null;
@@ -80,19 +69,19 @@ public class StandAloneValidatorFetcher implements IValidatorResourceFetcher {
     if (base == null) {
       return !url.startsWith("http://hl7.org/fhir") && !type.equals("canonical");
     }
-    
+
     if (base.equals("http://terminology.hl7.org")) {
-      pid = "hl7.terminology"; 
+      pid = "hl7.terminology";
     } else if (url.startsWith("http://hl7.org/fhir")) {
       pid = pcm.getPackageId(base);
-    } else { 
+    } else {
       pid = pcm.findCanonicalInLocalCache(base);
     }
-    ver = url.contains("|") ? url.substring(url.indexOf("|")+1) : null;
+    ver = url.contains("|") ? url.substring(url.indexOf("|") + 1) : null;
     if (pid == null && Utilities.startsWithInList(url, "http://hl7.org/fhir", "http://terminology.hl7.org")) {
       return false;
     }
-    
+
     if (url.startsWith("http://hl7.org/fhir")) {
       // first possibility: it's a reference to a version specific URL http://hl7.org/fhir/X.X/...
       VersionURLInfo vu = VersionUtilities.parseVersionUrl(url);
@@ -110,9 +99,8 @@ public class StandAloneValidatorFetcher implements IValidatorResourceFetcher {
         return pi.hasCanonical(url);
       }
     }
-    
 
- // we don't bother with urls outside fhir space in the standalone validator - we assume they are valid
+    // we don't bother with urls outside fhir space in the standalone validator - we assume they are valid
     return !url.startsWith("http://hl7.org/fhir") && !type.equals("canonical");
   }
 
@@ -122,7 +110,7 @@ public class StandAloneValidatorFetcher implements IValidatorResourceFetcher {
       try {
         json = JsonTrackingParser.fetchJson("http://hl7.org/fhir/mappingspaces.json");
         for (JsonObject ms : JSONUtil.objects(json, "spaces")) {
-          mappingsUris.add(JSONUtil.str(ms, "url"));       
+          mappingsUris.add(JSONUtil.str(ms, "url"));
         }
       } catch (IOException e) {
         // frozen R4 list
@@ -177,7 +165,7 @@ public class StandAloneValidatorFetcher implements IValidatorResourceFetcher {
 
   private String findBaseUrl(String url) {
     String[] p = url.split("\\/");
-    for (int i = 1; i< p.length; i++) {
+    for (int i = 1; i < p.length; i++) {
       if (Utilities.existsInList(p[i], context.getResourceNames())) {
         StringBuilder b = new StringBuilder(p[0]);
         for (int j = 1; j < i; j++) {
@@ -192,7 +180,7 @@ public class StandAloneValidatorFetcher implements IValidatorResourceFetcher {
 
   @Override
   public byte[] fetchRaw(String url) throws MalformedURLException, IOException {
-    throw new FHIRException("The URL '"+url+"' is not known to the FHIR validator, and has not been provided as part of the setup / parameters");
+    throw new FHIRException("The URL '" + url + "' is not known to the FHIR validator, and has not been provided as part of the setup / parameters");
   }
 
   @Override
@@ -207,20 +195,19 @@ public class StandAloneValidatorFetcher implements IValidatorResourceFetcher {
     if (root != null) {
       TerminologyClient c;
       c = TerminologyClientFactory.makeClient(root, context.getVersion());
-      return c.read(p[p.length-2], p[p.length-1]);
+      return c.read(p[p.length - 2], p[p.length - 1]);
     } else {
-      throw new FHIRException("The URL '"+url+"' is not known to the FHIR validator, and has not been provided as part of the setup / parameters");
+      throw new FHIRException("The URL '" + url + "' is not known to the FHIR validator, and has not been provided as part of the setup / parameters");
     }
   }
 
   private String getRoot(String[] p, String url) {
-    if (p.length > 3 && Utilities.isValidId(p[p.length-1]) && context.getResourceNames().contains(p[p.length-2])) {
+    if (p.length > 3 && Utilities.isValidId(p[p.length - 1]) && context.getResourceNames().contains(p[p.length - 2])) {
       url = url.substring(0, url.lastIndexOf("/"));
       return url.substring(0, url.lastIndexOf("/"));
     } else {
       return null;
     }
-    
   }
 
   @Override
